pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-document-cli"
version = "0.1.0"
description = "A CLI tool to generate and format professional PDFs using Gemini AI."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-genai",
    "rich",
    "markdown",
    "weasyprint",
    "python-dotenv"
]

# This is the magic part! It creates the terminal command.
[project.scripts]
ai-doc = "ai_doc.main:main"